readers: add partial_cancel function and implementation for rv1
Problem: there are no reader functions for partial job cancellation.

Add not-yet-implemented functions for all readers, and a full
implementation for RV1exec. For RV1exec, add the capability to unpack
and read the ranks to be cancelled from the free R payload. Also add a
modify_data_t struct to store the type of modification to be performed
on the jobid as well as ranks and types removed. Return the updated
modify_data_t struct to the traverser.