planner_multi: get random_access index based on resource type
Problem: a partial cancellation will accumulate resource counts to be
removed in a reader and the traverser. There is no guarantee that the
ordering of resource types encountered will be the same as when the
aggregate filter planner_multi was initialized. Therefore the vector
index of the resource does not provide a reliable mapping between
planner_multi resources and those found during the string unpacking and
traversal.

Add a function to get the multi_index index of a resource type based on
the resource name string.