planner_multi interface: add span partial cancel
Problem: partial cancellation requires the capability to remove parts of
a planner_multi span. The functionality takes a vector of type names
and new counts and must reduce the appropriate planner sub-spans.

Add the interface function to provide the require capability. The
function iterates through each resource type to be adjusted and maps
the name to its index in the planner_multi span. That identifies the
sub-planner span index for reduction of the planner resource type. The
loop then calls the planner_reduce_span function which performs the
resource reduction and returns whether the reduction resulted in a
complete removal of the span.

The return code of the function could be -1 if either the reduction
failed, or if the span no longer exists because it was removed by a
previous invocation of planner)reduce_span. Track the removed planner
sub-spans by -1 entries.

The second loop is required to check for resources not covered by the
resource_types function input vector. This can occur when a partial
reduction call does not include zero resource spans entered when the
planner_multi span was originally created. This can happen when a job
requests cores, but the planner_multi agfilter tracks cores and memory.
A span will be created for cores and memory, but the memory request
will be 0. We need to remove these spans.

Finally, check to see if all subspans were removed by counting the
number of -1 span entries.