test: add unreservable job test
problem: ensure we don't end up with unreservable jobs treated as
unsatisfiable again

solution: add a test that runs three jobs that are only allowed to run
on the same resource, and each take all remaining time.  This forces
jobs two and three to be unreservable until one completes.  It's
technically sensitive to timing, but in this constrained case I would be
surprised if it proves to be flaky.  Still should set something up we
can use in sharness for setting up jobs that wait on a trigger on a
domain socket so we can easily make this kind of test fully
deterministic.