formatting: add clang-format config and enable ci
problem: we end up spending way too much time worrying about manual code
layout

solution: automate it

Less glibly, this adds the clang-format file we've been working on for
core, but using the new standard line-length limit from the linux
kernel. It also switches us to using pre-commit to run our python
linting because that's an easy way to also apply clang-format in CI.
This _will_ fail when the checks run until other work is done, but this
is everything to run it and test it.