qmanager policies: add partial cancel in base, derived policies
Problem: the qmanager base policy and derived-classes do not have
partial cancellation functionality.

Add a virtual remove function overload in base, and overrides in FCFS
and backfill policies. In each case the policy needs to call the REAPI
module partial cancel function, which means it can't reside in the base
class.

Consolidate logic in the `remove` function to call partial cancel and
check if it fully removed the job's resources. If so, remove the job
from the allocated, running maps and set the state to be `COMPLETE`. Do
not enter the job into the completed map, because that will get popped
and cancelled again in `cancel_completed_jobs`. Note that the sched
loop needs to be cancelled and blocked jobs need to be reconsidered.
Finally, resume the sched loop to continue scheduling jobs.