t/qmanager-optimize: add comments and fix easy-backfill
problem: The easy-backfill portion of t1004-qmanager-optimize.t relies
on _exactly one_ scheduling loop pass to happen after a job is
cancelled.  Now that we re-enter the loop if something has changed
state, we've reached queue depth, and not reached the end of the pending
queue that no longer holds.  That test causes the loop to execute twice,
backfilling one more job.

solution: adjust the running count to 3. To improve things for next
time, add many more comments, and name the jobs by their numbers in the
script to make it easier to trace.